chest_xray/requirements.txt
streamlit
tflite-runtime==2.14.0
Pillow
numpy==1.24.4
